requirements.txt
# Essential packages
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
openai>=1.0.0

# NLP packages for full implementation
spacy>=3.7.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
yake>=0.4.8
keybert>=0.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Additional APIs and scraping
google-api-python-client>=2.100.0
google-generativeai>=0.8.0
selenium>=4.15.0
fake-useragent>=1.4.0

# Dev tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0